+++ docker/logs/nginx/.gitignore
@@ -1,0 +1,2 @@
+!symfony_access.log
+!symfony_error.log
