+++ intellst-api/composer.json
@@ -6,6 +6,7 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "lexik/jwt-authentication-bundle": "^2.8",
+        "nelmio/cors-bundle": "^2.0",
         "symfony/console": "5.1.*",
         "symfony/dotenv": "5.1.*",
         "symfony/flex": "^1.3.1",
