+++ intellst-api/config/routes.yaml
@@ -1,4 +1,4 @@
 
 login_check:
-  path:     /login_check
+  path:     /api/login_check
   methods:  [POST]
