+++ intellst-api/src/Service/Register.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Service;
+
+use App\Entity\User;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+
+class Register
+{
+    public function register(Request $request, UserPasswordEncoderInterface $encoder) : string
+    {
+        $em = $this->getDoctrine()->getManager();
+
+        $username = $request->request->get('username');
+        $password = $request->request->get('password');
+
+        $user = new User($username);
+        $user->setPassword($encoder->encodePassword($user, $password));
+        $em->persist($user);
+        $em->flush();
+
+        return new Response(sprintf('User %s successfully created', $user->getUsername()));
+    }
+}
